+++ lib/biotcm/databases.rb
@@ -4,4 +4,5 @@
 # directly or indirectly
 module BioTCM::Databases
   autoload(:HGNC, "biotcm/databases/hgnc")
+  autoload(:Cipher, "biotcm/databases/cipher")
 end